Make sleep use int and consolidate CRAM_SLEEP
Change Student::sleep signature from double to int (header and implementation) and remove an unnecessary static_cast when accumulating hoursSlept. Consolidate two CRAM_SLEEP constants into a single double CRAM_SLEEP and use it in cramForExam to decrement hoursSlept. Small type/constant cleanup and minor refactor to keep sleep handling consistent.

+++ Student.hpp
@@ -15,7 +15,7 @@ public:
     void drinkCoffee();
     void addAssignment();
     void submitAssignment();
-    void sleep(double hours);
+    void sleep(int hours);
     void cramForExam();
 
     std::string getName()           const;
